services/content-generator/requirements.txt
# Content Generator Service Requirements

fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aio-pika>=9.3.0
prometheus-client>=0.19.0

# OpenAI for content generation
openai>=1.3.0

# LLM Gateway - Additional providers (optional)
anthropic>=0.18.0
google-generativeai>=0.3.0

# Celery for scheduled tasks
celery>=5.3.0
redis>=5.0.0

# HTTP client for image APIs
httpx>=0.25.0

# MinIO/S3 for image storage
boto3>=1.34.0
aioboto3>=12.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0